Palindrome Task
Let's call a string interesting if there is a substring of the length of half of the string or more, that reads the same forwards and backwards.
For example, 'abc' is not interesting, however, 'aabc' is interesting since the substring 'aa' is a palindrome. If a string is interesting, return the substring that makes that possible. Otherwise, simply return False.
Ignore capitalization and return None for empty string. 

Examples: 
>>> is_interesting('') == None 
>>> is_interesting('abc') == False 
>>> is_interesting('aabc') == aa